dom/media/webaudio/test/test_nodeCreationDocumentGone.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test whether we can create an AudioContext interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.requestCompleteLog();
SimpleTest.waitForExplicitFinish();


SpecialPowers.pushPrefEnv({
  set: [
    [ "dom.disable_open_during_load", false ]
  ]
}).then(function() {
  info("after pref");
  var sub = encodeURI("data:text/html,<!DOCTYPE html>\n"+
                      "<html><script>"+
                      "var context = new AudioContext();"+
                      "setTimeout(function(){window.close();},1000);\x3C/script></html>");
  window.onload = function(){
    info("after onload");
    var a = window.open(sub);
    info("after open: " + a);
    a.onbeforeunload = function(){
      setTimeout(function(){
        try {
          info("before sp");
          a.context.createScriptProcessor(512, 1, 1);
          info("after sp");
        } catch(e) {
          ok (true,"got exception");
        }
        setTimeout(function() {
          info("finish");
          ok (true,"no crash");
          SimpleTest.finish();
        }, 0);
      }, 0);
    };
  };
});


</script>
</pre>
</body>
</html>
